12/05 (김호식) 19:50 조명추가및 중간커밋

+++ user/src/main/java/edu/sm/app/dto/AvgTHDto.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AvgTHDto {
     private Double avgTemperature;
-    private Double avgHumidity;
+    private int avgHumidity;
 }
